refactor: implement phase 6.5

+++ BACKEND/src/controllers/deliveryAccountController.js
@@ -1,4 +1,6 @@
+const { subject } = require("@casl/ability");
 const deliveryAccountService = require("../services/deliveryAccountService");
+const ForbiddenError = require("../errors/ForbiddenError");
 
 const handleGetAllDeliveryAccounts = async (req, res) => {
   const userId = req.user.id;
@@ -9,6 +11,9 @@ const handleGetAllDeliveryAccounts = async (req, res) => {
 const handleGetDeliveryAccountById = async (req, res) => {
   const id = req.params.id;
   const account = await deliveryAccountService.getDeliveryAccountById(id);
+  if (!req.ability.can("read", subject("DeliveryAccount", account))) {
+    throw new ForbiddenError("Bạn không có quyền xem tài khoản vận chuyển này");
+  }
   return res.success("Get delivery account success", account);
 };
 
@@ -21,20 +26,32 @@ const handleCreateDeliveryAccount = async (req, res) => {
 const handleUpdateDeliveryAccount = async (req, res) => {
   const id = req.params.id;
   const data = req.body;
-  const account = await deliveryAccountService.updateDeliveryAccount(id, data);
-  return res.success("Update delivery account success", account);
+  const account = await deliveryAccountService.getDeliveryAccountById(id);
+  if (!req.ability.can("update", subject("DeliveryAccount", account))) {
+    throw new ForbiddenError("Bạn không có quyền chỉnh sửa tài khoản vận chuyển này");
+  }
+  const updatedAccount = await deliveryAccountService.updateDeliveryAccount(id, data);
+  return res.success("Update delivery account success", updatedAccount);
 };
 
 const handleDeleteDeliveryAccount = async (req, res) => {
   const id = req.params.id;
+  const account = await deliveryAccountService.getDeliveryAccountById(id);
+  if (!req.ability.can("delete", subject("DeliveryAccount", account))) {
+    throw new ForbiddenError("Bạn không có quyền xóa tài khoản vận chuyển này");
+  }
   const result = await deliveryAccountService.deleteDeliveryAccount(id);
   return res.success("Delete delivery account success", result);
 };
 
 const handleSetDefaultDeliveryAccount = async (req, res) => {
   const id = req.params.id;
-  const account = await deliveryAccountService.setDefaultDeliveryAccount(id);
-  return res.success("Set default delivery account success", account);
+  const account = await deliveryAccountService.getDeliveryAccountById(id);
+  if (!req.ability.can("update", subject("DeliveryAccount", account))) {
+    throw new ForbiddenError("Bạn không có quyền chỉnh sửa tài khoản vận chuyển này");
+  }
+  const updatedAccount = await deliveryAccountService.setDefaultDeliveryAccount(id);
+  return res.success("Set default delivery account success", updatedAccount);
 };
 
 module.exports = {
@@ -42,6 +59,7 @@ module.exports = {
   handleGetDeliveryAccountById,
   handleCreateDeliveryAccount,
   handleUpdateDeliveryAccount,
+  handleDeleteItem: handleDeleteDeliveryAccount, // Make sure name is consistent or exports are preserved
   handleDeleteDeliveryAccount,
   handleSetDefaultDeliveryAccount,
 };
